pyproject.toml
[tool.poetry]
name = "wagtail-cfran"
version = "1.4.3"
description = "Gestionnaire de contenu permettant de créer et gérer un site internet basé sur le Système de design de l'État, accessible et responsive"
authors = [
  "Sébastien Reuiller <sebastien.reuiller@beta.gouv.fr>",
  "Sylvain Boissel <sylvain.boissel@beta.gouv.fr>",
]
maintainers = [
  "Sylvain Boissel <sylvain.boissel@beta.gouv.fr>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
wagtail = "^6.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
gunicorn = "^22.0.0"
django-sass-processor = "^1.3"
libsass = "^0.22.0"
dj-static = "^0.0.6"
wagtailmenus = "^4.0"
wagtail-modeladmin = "^2.0.0"
wagtail-markdown = "^0.11.1"
unidecode = "^1.3.8"
django-storages = {extras = ["s3"], version = "^1.14.2"}
boto3 = "^1.34.56"
beautifulsoup4 = "^4.12.3"
django-taggit = "^5.0.1"
wagtail-localize = "^1.9"
django-cfran = "^11.2.5"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"
djlint = "^1.34.0"
django-extensions = "^3.2.3"
ipython = "^8.18.1"
libsass = "^0.22.0"
django-compressor = "^4.4"
faker = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 119

[tool.pyright]
reportAttributeAccessIssue = false
reportArgumentType = false
reportFunctionMemberAccess = false
reportGeneralTypeIssues = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportOptionalCall = false
reportOptionalMemberAccess = false
reportUnboundVariable = false
